+++ flask_multipass/core.py
@@ -155,7 +155,13 @@ class Multipass:
         # Browsers treat backslashes like forward slashes, while urllib doesn't.
         # Since we just want to validate scheme and netloc here, we normalize
         # slashes to those recognized by urllib.
-        url_info = urlsplit(url.replace('\\', '/'))
+        url = url.replace('\\', '/')
+        # Browsers are great at accepting crappy data, so `////example.com` is
+        # handled like `//example.com`, even though it's invalid and `urlparse`
+        # will not set a netloc when receiving such a URL.
+        if url.startswith('///'):
+            return False
+        url_info = urlsplit(url)
         if url_info.scheme and url_info.scheme not in {'http', 'https'}:
             return False
         if url_info.scheme and not url_info.netloc: